+++ vEngine/vEngine/Engine/Source/Model.cpp
@@ -254,7 +254,7 @@ namespace vEngine
 					textures_.push_back(tex);
 					tex_srvs_.push_back(Context::Instance().GetRenderFactory().MakeRenderBuffer(textures_.back(), AT_GPU_READ_WRITE,BU_SHADER_RES));
 					//index 0 reserved for null
-					mat->diffuse_tex = static_cast<float>(textures_.size());
+					mat->diffuse_tex = (uint32_t)textures_.size();
 				}
 			}
 			else
@@ -272,7 +272,7 @@ namespace vEngine
 					textures_.push_back(tex);
 					tex_srvs_.push_back(Context::Instance().GetRenderFactory().MakeRenderBuffer(textures_.back(), AT_GPU_READ_WRITE,BU_SHADER_RES));
 					//index 0 reserved for null
-					mat->normalmap_tex = static_cast<float>(textures_.size());
+					mat->normalmap_tex = (uint32_t)textures_.size();
 				}
 			}
 			else
